requirements.txt
Flask==3.0.2
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0 